Refactor/split api into separate files; easily identify API calls
